Moves checkAgeSex(Date dob, Sex sex, Title title) method from CommonFunctions to Person for proper encapsulation of responsibilities
